requirements.txt
typer~=0.12.5
fastapi~=0.112.2
click~=8.1.7
setuptools~=73.0.1
motor~=3.6.0
pytest~=8.3.3
pymongo~=4.9.2
pydantic~=2.8.2
python-dotenv~=1.0.1
makefast~=0.1
mysql-connector-python~=9.0.0
starlette~=0.38.2